batch_jobs_to_bigquery/scrape_triggered_gcs_to_bq/TwitterPulseApp/requirements.txt

flask==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

spacy>=3.0.0,<4.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz#egg=en_core_web_sm

spacytextblob==3.0.1

tweepy==4.4.0

vadersentiment==3.3.2

google-cloud-secret-manager==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

google-cloud-storage==1.44.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

gunicorn==20.1.0

pandas==1.3.5